+++ Verizon/CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Changelog
+  * 1.9.0.0
+    * This version of the adapters has been certified with Verizon 1.9.0 and MoPub 5.15.0.
+    * Refactor native ad impression tracking logic. No external changes for publishers.
+    * Native ads now require a disclaimer/sponsored text view.
+  
   * 1.8.2.1
     * Fix a bug where the rewarded video adapter fails to request a new ad after a show-related error happens.